Updated designs for the unit page (#1325)
* Design updates

* Refactor out UnitCard. Apply styles for desktop

* Mobile breakpoint for logo container

* Migration for updated copy

* Remove log line, fix migration

* Apply loading styles
