+++ frontends/mit-open/src/pages/UnitsListingPage/UnitsListingPage.tsx
@@ -1,26 +1,24 @@
+import React from "react"
 import {
   useLearningResourcesSearch,
   useOfferorsList,
 } from "api/hooks/learningResources"
 import {
   Banner,
-  Card,
   Container,
-  Skeleton,
   Typography,
   styled,
   Breadcrumbs,
+  theme,
 } from "ol-components"
 import { RiBookOpenLine, RiSuitcaseLine } from "@remixicon/react"
-import React from "react"
 import {
   LearningResourceOfferorDetail,
   LearningResourcesSearchResponse,
-  OfferedByEnum,
 } from "api"
 import { MetaTags } from "ol-utilities"
-import { useChannelDetail } from "api/hooks/channels"
 import { HOME } from "@/common/urls"
+import { UnitCards, UnitCardLoading } from "./UnitCard"
 
 const UNITS_BANNER_IMAGE = "/static/images/background_steps.jpeg"
 const DESKTOP_WIDTH = "1056px"
@@ -61,7 +59,7 @@ const PageContent = styled.div(({ theme }) => ({
   flexDirection: "column",
   alignItems: "center",
   padding: "40px 10px 80px 10px",
-  gap: "80px",
+  gap: "48px",
   [theme.breakpoints.down("md")]: {
     padding: "40px 0px 30px 0px",
     gap: "40px",
@@ -84,16 +82,21 @@ const PageHeaderContainerInner = styled.div({
   display: "flex",
   flexDirection: "column",
   maxWidth: "1000px",
+  border: `1px solid ${theme.custom.colors.lightGray2}`,
+  backgroundColor: theme.custom.colors.white,
+  borderRadius: "8px",
+  padding: "32px",
+  [theme.breakpoints.down("md")]: {
+    backgroundColor: "transparent",
+    border: "none",
+    padding: "0",
+  },
 })
 
 const PageHeaderText = styled(Typography)(({ theme }) => ({
   color: theme.custom.colors.black,
   ...theme.typography.subtitle1,
 }))
-
-const CardStyled = styled(Card)({
-  height: "100%",
-})
 
 const UnitContainer = styled.div(({ theme }) => ({
   display: "flex",
@@ -146,76 +149,11 @@ const GridContainer = styled.div(({ theme }) => ({
   display: "grid",
   gap: "25px",
   gridTemplateColumns: "repeat(2, 1fr)",
+  width: "100%",
   [theme.breakpoints.down("md")]: {
     gridTemplateColumns: "1fr",
   },
 }))
-
-const UnitCardContainer = styled.div({
-  display: "flex",
-  flexDirection: "column",
-  alignItems: "center",
-  padding: "16px",
-  height: "100%",
-})
-
-const UnitCardContent = styled.div({
-  display: "flex",
-  flexDirection: "column",
-  flexGrow: 1,
-})
-
-const LogoContainer = styled.div({
-  display: "flex",
-  justifyContent: "center",
-  alignItems: "center",
-  height: "128px",
-})
-
-const UnitLogo = styled.img({
-  display: "flex",
-  flexDirection: "column",
-  justifyContent: "center",
-  alignItems: "center",
-  height: "50px",
-  maxWidth: "100%",
-})
-
-const ValuePropContainer = styled.div({
-  display: "flex",
-  flexDirection: "column",
-  alignItems: "flex-start",
-  justifyContent: "flex-start",
-  flexGrow: 1,
-  paddingBottom: "16px",
-})
-
-const ValuePropText = styled(Typography)(({ theme }) => ({
-  alignSelf: "stretch",
-  color: theme.custom.colors.darkGray2,
-  ...theme.typography.body2,
-}))
-
-const CountsTextContainer = styled.div({
-  display: "flex",
-  justifyContent: "flex-end",
-  gap: "10px",
-})
-
-const CountsText = styled(Typography)(({ theme }) => ({
-  textAlign: "center",
-  color: theme.custom.colors.silverGrayDark,
-  ...theme.typography.body3,
-}))
-
-const unitLogos = {
-  [OfferedByEnum.Mitx]: "/static/images/unit_logos/mitx.svg",
-  [OfferedByEnum.Ocw]: "/static/images/unit_logos/ocw.svg",
-  [OfferedByEnum.Bootcamps]: "/static/images/unit_logos/bootcamps.svg",
-  [OfferedByEnum.Xpro]: "/static/images/unit_logos/xpro.svg",
-  [OfferedByEnum.Mitpe]: "/static/images/unit_logos/mitpe.svg",
-  [OfferedByEnum.See]: "/static/images/unit_logos/see.svg",
-}
 
 interface UnitSectionProps {
   id: string
@@ -264,95 +202,6 @@ const UnitSection: React.FC<UnitSectionProps> = (props) => {
         )}
       </GridContainer>
     </UnitContainer>
-  )
-}
-
-interface UnitCardsProps {
-  units: LearningResourceOfferorDetail[] | undefined
-  courseCounts: Record<string, number>
-  programCounts: Record<string, number>
-}
-
-const UnitCards: React.FC<UnitCardsProps> = (props) => {
-  const { units, courseCounts, programCounts } = props
-  return (
-    <>
-      {units?.map((unit) => {
-        const courseCount = courseCounts[unit.code] || 0
-        const programCount = programCounts[unit.code] || 0
-        const logo =
-          unitLogos[unit.code as OfferedByEnum] ||
-          `/static/images/unit_logos/${unit.code}.svg`
-        return unit.value_prop ? (
-          <UnitCard
-            key={unit.code}
-            unit={unit}
-            logo={logo}
-            courseCount={courseCount}
-            programCount={programCount}
-          />
-        ) : null
-      })}
-    </>
-  )
-}
-
-interface UnitCardProps {
-  unit: LearningResourceOfferorDetail
-  logo: string
-  courseCount: number
-  programCount: number
-}
-
-const UnitCard: React.FC<UnitCardProps> = (props) => {
-  const { unit, logo, courseCount, programCount } = props
-  const channelDetailQuery = useChannelDetail("unit", unit.code)
-  const channelDetail = channelDetailQuery.data
-  const unitUrl = channelDetail?.channel_url
-  return channelDetailQuery.isLoading ? (
-    <UnitCardLoading />
-  ) : (
-    <CardStyled href={unitUrl}>
-      <Card.Content>
-        <UnitCardContainer>
-          <UnitCardContent>
-            <LogoContainer>
-              <UnitLogo src={logo} alt={unit.name} />
-            </LogoContainer>
-            <ValuePropContainer>
-              <ValuePropText>{unit.value_prop}</ValuePropText>
-            </ValuePropContainer>
-            <CountsTextContainer>
-              <CountsText data-testid={`course-count-${unit.code}`}>
-                {courseCount > 0 ? `Courses: ${courseCount}` : ""}
-              </CountsText>
-              <CountsText data-testid={`program-count-${unit.code}`}>
-                {programCount > 0 ? `Programs: ${programCount}` : ""}
-              </CountsText>
-            </CountsTextContainer>
-          </UnitCardContent>
-        </UnitCardContainer>
-      </Card.Content>
-    </CardStyled>
-  )
-}
-
-const UnitCardLoading = () => {
-  return (
-    <Card>
-      <Card.Content>
-        <UnitCardContainer>
-          <UnitCardContent>
-            <LogoContainer>
-              <Skeleton variant="rectangular" width={500} height={50} />
-            </LogoContainer>
-            <ValuePropContainer>
-              <Skeleton variant="text" width={500} height={200} />
-            </ValuePropContainer>
-          </UnitCardContent>
-        </UnitCardContainer>
-      </Card.Content>
-    </Card>
   )
 }
